+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.8.0 - In Development
 
+* [#169](https://github.com/adamcaudill/yawast/issues/169) - Bug: Error in connecting to SSL Labs
+
 ## 0.7.1 - 2019-05-07
 
 * [#37](https://github.com/adamcaudill/yawast/issues/37) - Batch Scanning Mode